12.92 GBP/pack of 20 @ Waitrose online 2020-10-30
Based on £64.60 for 100s pack

Fixed tests to use the currency formatter rather than construct the
expected string directly - it became too unwieldy once the examples
passed £1,000 and would need a comma adding in.

This also fixes the test that was failing when the price was £12.00 -
it was the test constructing its expected result that was wrong, not
the code!