refactor: begin extracting savings details
Rather than having specific savings details (price, packs, and so on)
hard-coded inside the calculation function, begin extracting them, so
that we can move towards being able to calculate the answer for many
people, not just me! So that the code and site can be more useful.

For the first change, use a maker to recreate the original function by
capturing various arguments. This moves the details out, while leaving
the test code unchanged for the moment.