feat: show (broken) "Celebrations" on main page not console
Instead of logging to the browser's console, show the - currently sort
of broken - Celebrations on the main HTML page.
All it shows at the moment is the number of weeks, with an incrementing
fraction. As it happens, this at least updates very frequently, so the
page looks like it's alive: it updates more quickly than the Savings
value, which ticks over more slowly.