feat: show (broken) "Summary" on main page not console
Instead of logging to the browser's console, show the - currently sort
of broken - Summary on the main HTML page.
At the moment it shows a broken "time" difference,
a week with a forced (0 days) that contrasts oddly with the Celebration
section that shows week's fraction, and the number of days.
Just a work in progress. Better on the page than in the console, which
still has a couple of lines of info that need reconsidering.